fix: address PR review — float precision for bar color, periodic cache cleanup
- Use float instead of int for bar color threshold check — 85.5%
  now correctly shows red instead of truncating to 85 (yellow)
- Add periodic cache cleanup: removes files older than 2 days from
  user-scoped cache dir (runs at most once per hour on cache write)

+++ claude_statusline/bar.py
@@ -35,7 +35,7 @@ def render_bar(pct, width=20, theme=None, compaction_threshold=None):
         return ""
 
     # Use raw percentage for color (reflects actual context fill)
-    color = _bar_color(max(0, min(100, int(pct))))
+    color = _bar_color(max(0, min(100, float(pct))))
 
     # Scale fill width relative to compaction threshold if configured
     if compaction_threshold and 0 < compaction_threshold <= 100:
